Added encoders
Created an AutoDrive class to store all the actions during autonomous. For now it only goes straight a certain distance.
Created a ResetEncoders to reset the encoders. This is only for debugging and will likely be disabled during competition.
Edited the get average right and left distance for the encoders in Drivetrain in order for it to get the total distance instead of the difference in distance (due to some of the motors going the opposite way)