Commit with New Pivot and Extension
This is a new commit with a new command called pivotToAngle which allows us to change our angle at our own convenience. The current angles are 30 deg with button 10 and 90 deg with button 11. In addition to this, the extension now has the encoder value functions so that the arm can go according to the encoder values or manually. --Sagar